Add rm -rf flag

+++ src/client/commands/rm.js
@@ -1,14 +1,29 @@
 import {
     isNil,
     find,
-    propEq
+    propEq,
+    isEmpty,
+    contains,
+    length,
+    map
 } from 'ramda';
 
 import { store } from '../index';
 import { getFiles } from './ls';
-import { removeFile } from '../actions/fileSys';
+import { removeFile, removeFolder } from '../actions/fileSys';
 
-const rm = (termId, name) => {
+const checkFlags = (flags, supportedFlags) => {
+    let isFlagsSupported = true;
+
+    map(flag => {
+        if(!contains(flag, supportedFlags))
+            isFlagsSupported = false
+    },flags);
+    console.log('isFlagsSupported: ', isFlagsSupported)
+    return isFlagsSupported ? 0 : -1;
+}
+
+const rm = (termId, name, flags, supportedFlags) => {
     const state = store.getState();
     const { terms, fileSys } = store.getState();
     const term = find(propEq('id', termId))(terms);
@@ -16,17 +31,27 @@ const rm = (termId, name) => {
     const files = getFiles(path, fileSys);
     const isFileExisting = !isNil(find(propEq('value', name))(files));
     
+    if(checkFlags(flags, supportedFlags) === -1)
+        return [{
+            value: `rm: unknow flag`,
+            color: 'white',
+        }];
     if(isNil(name))
         return [{
-            value: 'usage: rm [-f] [-R] file ...',
+            value: 'usage: rm [-rf] file ...',
             color: 'white',
         }];
-    if(!isFileExisting)
-        return [{
-            value: `rm: ${name}: File doesn't exists`,
-            color: 'white',
-        }];
-    store.dispatch(removeFile(path, name));
+    if(isEmpty(flags)) {
+        if(!isFileExisting)
+            return [{
+                value: `rm: ${name}: File doesn't exists`,
+                color: 'white',
+            }];
+        store.dispatch(removeFile(path, name));
+    }
+    else if (contains("-rf", flags)) {
+        store.dispatch(removeFolder(path, name));
+    }
     return;
 };
